requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-google-genai
sentence-transformers
langchain-community
numpy==1.26.4
pypdf
requests